backend/requirements.txt
# Framework Web
Flask==3.0.3
Flask-CORS==6.0.0

# Servidor de Produção
gunicorn==22.0.0

# Bibliotecas de AI/ML
google-generativeai==0.8.0

# Processamento de PDF
pypdf==6.0.0

# Processamento de Email
email-validator==2.2.0

# Configuração e Variáveis de Ambiente
python-dotenv==1.0.1

# Requisições HTTP
requests==2.32.4

# Ferramentas de Desenvolvimento
pytest==8.3.3
pytest-flask==1.3.0
